Reorganize utilities file.
