+++ app/assets/scripts/utils.js
@@ -1,0 +1,10 @@
+'use strict';
+
+/**
+ * Delays the execution in x milliseconds.
+ *
+ * @param {int} millis Milliseconds
+ */
+export function delay (millis) {
+  return new Promise(resolve => setTimeout(resolve, millis));
+}
